requirements.txt
luma.oled
RPi.GPIO
smbus2
psutil
sphinx
